Można edytować i kończyć tylko swoją aukcję

+++ src/main/java/com/javabiz/auctionsite/service/auctionmanagement/AuctionController.java
@@ -1,12 +1,15 @@
 package com.javabiz.auctionsite.service.auctionmanagement;
 
+import com.javabiz.auctionsite.Security.UserDetailsServiceImplementation;
 import com.javabiz.auctionsite.service.commons.OfferDto;
 import com.javabiz.auctionsite.service.mailservice.MailService;
 import com.javabiz.auctionsite.service.model.Auction;
 import com.javabiz.auctionsite.service.commons.AuctionDto;
 import com.javabiz.auctionsite.service.model.Offer;
+import com.javabiz.auctionsite.service.model.UserModel;
 import com.javabiz.auctionsite.service.usermanagement.UserRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -15,6 +18,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @AllArgsConstructor
@@ -23,6 +27,7 @@ public class AuctionController {
 
     private final AuctionService auctionService;
     private final MailService mailService;
+    private final UserRepository userRepository;
 
     @GetMapping()
     public String getAll(Model model) {
@@ -52,17 +57,22 @@ public class AuctionController {
     }
 
     @GetMapping("/{id}/offers/{offerId}")
-    public String endAuctionByChoosingTheOffer(@PathVariable Long id, @PathVariable Long offerId){
+    public String endAuctionByChoosingTheOffer(@PathVariable Long id, @PathVariable Long offerId, Authentication authentication){
         Auction auction = auctionService.findAuction(id);
         Offer offer = auctionService.findOffer(offerId);
 
         assert auction.getOwner()!=null;
         assert offer.getOffering()!=null;
 
-        auctionService.endAuction(id, offerId);
-        mailService.sendConfirmationMailToSeller(auction.getOwner());
-        mailService.sendConfirmationMailToBuyer(offer.getOffering());
-        return "redirect:/auction/";
+        if(!Objects.equals(auction.getOwner().getUsername(), authentication.getName())) {
+            return "redirect:/auction/";
+        }
+        else {
+            auctionService.endAuction(id, offerId);
+            mailService.sendConfirmationMailToSeller(auction.getOwner());
+            mailService.sendConfirmationMailToBuyer(offer.getOffering());
+            return "redirect:/auction/";
+        }
     }
 
     @GetMapping("/{id}/offer")
@@ -77,7 +87,8 @@ public class AuctionController {
     @PostMapping("/makeOffer/{id}")
     public String addOffer(@Valid OfferDto offer,
                              @PathVariable Long id,
-                             RedirectAttributes redirectAttributes) {
+                             RedirectAttributes redirectAttributes, Authentication authentication) {
+        offer.setOffering(userRepository.findByUsername(authentication.getName()));
         offer.setAuction(auctionService.findAuction(id));
         auctionService.createOffer(offer);
         redirectAttributes.addFlashAttribute("success", "Record has been successfully added.");
@@ -94,17 +105,22 @@ public class AuctionController {
 
     @PostMapping("/add")
     public String addAuction(@Valid AuctionDto auction,
-                              RedirectAttributes redirectAttributes) {
+                             RedirectAttributes redirectAttributes, Authentication authentication) {
+        auction.setOwner(userRepository.findByUsername(authentication.getName()));
         auctionService.createAuction(auction);
         redirectAttributes.addFlashAttribute("success", "Record has been successfully added.");
         return "redirect:/auction/";
     }
 
     @GetMapping("/edit/{id}")
-    public String editAuction(@PathVariable Long id, Model model) {
+    public String editAuction(@PathVariable Long id, Model model, Authentication authentication) {
         Auction auction = auctionService.findAuction(id);
         AuctionDto auctionDto = new AuctionDto(auction.getTitle(), auction.getContent());
 
+        if(!Objects.equals(auction.getOwner().getUsername(), authentication.getName())){
+            return "redirect:/auction/";
+        }
+        else {
         if (!model.containsAttribute("auction")) {
             model.addAttribute("auction", new AuctionDto());
         }
@@ -112,6 +128,7 @@ public class AuctionController {
         model.addAttribute("auctionId", id);
         model.addAttribute("auction", auctionDto);
         return "auction/editAuction";
+        }
     }
 
     @PutMapping("/{id}/update")
@@ -132,8 +149,14 @@ public class AuctionController {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteAuction(@PathVariable("id") Long id) {
-        auctionService.deleteAuction(id);
-        return "redirect:/auction";
+    public String deleteAuction(@PathVariable("id") Long id, Authentication authentication) {
+        Auction auction = auctionService.findAuction(id);
+        if(!Objects.equals(auction.getOwner().getUsername(), authentication.getName())){
+            return "redirect:/auction/";
+        }
+        else {
+            auctionService.deleteAuction(id);
+            return "redirect:/auction";
+        }
     }
 }
